requirements.txt
line-bot-sdk
flask
python-firebase
requests
qrcode
pillow
imgurpython